backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
PyPDF2==3.0.1
boto3==1.28.85
mysql-connector-python==8.1.0